Reintroduce the protocol member and set_protocol since they are needed in the workflow and clarify their meaning. Adapt downloaders.
The protocol here is passed from the file (so it's "directftp" for
DirectFTPDownload). See DownloadService.get_handler for instance. Hence,
the real protocol to use (by cURL or other) must be handled separately.
There are other ways but this is probably the one with the lower impact
on other modules.

This essentially undo d92ccdc9e2d3851db3c564280465db22d041beb0 and
67a269516c67db1e09c5cf26451a02a16832d6c6.

Tests are unchanged but note that some errors persists in module biomaj.