Refactor downloaders (#7)
* Minor PEP8 corrections.

* Minor edits.

* Refactor the download method: most classes only define the method to
download one file.

All tests pass. The only one that was modified is
TestBiomajRSYNCDownload.test_rsync_download because it used a method
that was renamed and need a bit more information.

Note that LocalDownload doesn't work that way and IRODSDownload is still
to do.

* Centralize the modification of rfiles into a method instead of doing it in the interface (and other methods).

* Remove method FTPDownload.match since its implementation is almost identical to the generic one.

* Correctly set the list of files to download.

* Factor some parts of list method for FTP and HTTP downloaders.

* Correct comment.

* Report code from FTPDownload.match (this could cause bugs).

This was suggested by O. Sallou.

* Add a test of IRODSDownload.download (it requires a local iRODS server so we disable it in CI).

* Use the generic download method. Clean file.

* Remove unused imports.

* Remove unused import.

* Delete duplicate header_function (it already exists in the FTPDownload class).

* Use class name instead of protocol when logging.

* Remove unused imports.

* Transform protocol to a class constant.

* Set protocol for class localDownload.

* Add some comments and remove useless code.

* Correct test selection formula.

We want tests that are not `local_irods` and not `network`.

* Add some notes about tests.

* PEP8 edits to satisfy flake8.

* Another round of flake8 fixes.

* Typo.

* Raise exception if uncompress failed.

This should be enough to restart the download.

* Move back test of archives in FTPDownload for now because it's the only Downloader that retry in case of temporary failure.

* Comment.

* Remove useless import.

* Typo.

* Add UNCOMPRESS_SKIP_CHECK in environment if needed but don't remove it (because it breaks CLI).

* Correctly construct FTPDownload in test (8b1b167da3be3117c5ba3f19d16db1352c8b4d97 should not have changed that).

* Rename DirectHttpDwonload to DirectHTTPDownload.

This might introduce some problems in other packages but few are expected.

* Clear os.chdir() calls.

Don't chdir in interface (this was buggy and generally useless). We do
it in RSYNCDownloader constructor because it's needed here. Previously,
the user had to call set_offline_dir() before calling list() match().

* Reorganize imports.

* Merge FTPDownload and HTTPDownload into CurlDownload. This class manage several protocols (http, https, ftp, sftp). Adapt code all around.

This was done to re-introduce download from https URL.

This is not great from OOP point of view but the old FTPDownload class
already contained HTTP parts so this clarify the situation.

* Get the value of UNCOMPRESS_SKIP_CHECK in constructor.

* Remove useless code (should have been done in b33114d4afec219b89f1a84f08089aea6b0ead2d).

* Add test for SFTP download.

* Add an environment variable to bypass SSL host verification because it might fails.

The problem seems to come from the SSH backend used by cURL (libssh or libssh2).
Those backends can't read known_hosts files created by recent versions of OpenSSH.

* Use elif instead of if.

* Better error messages.

* Don't call set_protocol in DownloadService.get_handler since it has been removed (see 67a2695).

Note that I'm not sure if we should re-introduce it. Also note that this is not
detected by our tests unit tests but the tests in module:`biomaj`.

* Reintroduce case for DirectFTPS.

Should have been done in b24fe3d1723d293967e3ad2721d02b7101540ec4 (when
merging master).

* Use protocol name instead of number (as needed when constructing CurlDownloader).

This was revealed by tests in biomaj package.

* Reintroduce the protocol member and set_protocol since they are needed in the workflow and clarify their meaning. Adapt downloaders.

The protocol here is passed from the file (so it's "directftp" for
DirectFTPDownload). See DownloadService.get_handler for instance. Hence,
the real protocol to use (by cURL or other) must be handled separately.
There are other ways but this is probably the one with the lower impact
on other modules.

This essentially undo d92ccdc9e2d3851db3c564280465db22d041beb0 and
67a269516c67db1e09c5cf26451a02a16832d6c6.

Tests are unchanged but note that some errors persists in module biomaj.

* Use the real protocol when forging URLs.

* Remove useless comment.

* Typo.

* Use a variable for file URL in list.

* Misc.

* Close file before testing archive.

This allow TestBiomajHTTPSDownload.test_download to pass.

* Rename parameter.

* Remove __init__ since it's useless (everything is done in parent).

* Correct implementation of direct downloaders.

* Improve comments on direct downloaders.

* Check number of files to download in set_files_to_download instead of download. Update comments.

All tests in biomaj and biomaj-download pass.

* _download reuses the inner Curl object instead of creating a new one for each file. Adapt list in CurlDownloader and subclasses.

[cURL doc](https://curl.haxx.se/libcurl/c/curl_easy_perform.html) states
that using the perform method on the same object several times is
beneficial for performance.

* Use _file_url in loop (even if the loop is only for 1 element).

* Factorize configuration of cURL object into a method.

* Reorganize file.

* Reorganize file by grouping methods into 2 broad categories and add some comments.

* Override set_server in CurlDownloader since some members depend on it. Add doc.

* A bit of doc.

* Rename private method.

* Remove useless method override.

* PEP8 edits to satisfy flake8.

* Correct doc.